% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{begin}
\alias{begin}
\title{Begin a code block}
\usage{
begin(...)
}
\arguments{
\item{...}{any number of comma-separated R statements}
}
\value{
Last evaluated statement or NULL.
}
\description{
An alias to \code{{}} braces to allow multiple statements to
be expressed in a single call.
}
\examples{
begin(a <- 1, a + 1)  # 2
}
