% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{displayTree}
\alias{displayTree}
\title{Print a console tree}
\usage{
displayTree(root, nodes = function(x) x$nodes, label = function(x) x$label)
}
\arguments{
\item{root}{A tree, a data structure like a nested list with one "root", that is, a
single node with possibly multiple subnodes each with subnodes recursively.
Each node has a "label" string to be printed.}

\item{nodes}{Function that extract list of nodes from each node. By default, nodes are
the "nodes" element of a list.}

\item{label}{Function that extract label from each node. By default, stored in "label"
element of a list.}
}
\value{
NULL invisibly
}
\description{
Print a console tree
}
\examples{
tree.rpois <- function(lambda = 2, decr = 0.75) {
  cnt <- 0
  .f <- function(i, p) {
    cnt <<- cnt + 1
    list(label = sQuote(cnt),
         nodes = lapply(seq_len(rpois(1, p)), .f, p * decr))
  }
  .f(1, lambda)
}
set.seed(847)
tree <- tree.rpois()
displayTree(tree)

}
