% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_mandelbrot}
\alias{plot_mandelbrot}
\title{Plot the beautiful Mandelbrot set}
\usage{
plot_mandelbrot(xlim, ylim, res = 0.002, max_depth = 200, nclr = 20)
}
\arguments{
\item{xlim}{plot x limits}

\item{ylim}{plot y limits}

\item{res}{resolution}

\item{max_depth}{max depth of iterations to find if a point is in the set}

\item{nclr}{number of colors}
}
\value{
invisible()
}
\description{
Only an exapmle of the use of Tailcall to calculate a Mandelbrot point set.
}
\examples{
plot_mandelbrot(x = c(-2, 0.5), y = c(-1, 1), res= 0.02, nclr = 20)
plot_mandelbrot(x = c(-0.06, -0.05), y = c(-0.83, -0.82),
               res= 0.0005, max_depth = 2000)
}
