% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rebase}
\alias{rebase}
\alias{print.rebase}
\title{Calculate number in another base (radix)}
\usage{
rebase(n, base, prec)

\method{print}{rebase}(x, ...)
}
\arguments{
\item{n}{numeric vector to change base, or the result of a previous call to rebase.}

\item{base}{base to change representation to. If ommited and n is a result of rebase,
the return value is the numeric vector.}

\item{prec}{precision, in number of digits}

\item{x}{a \emph{rebase} object}

\item{...}{further arguments passed to base \code{print}}
}
\value{
for a numeric vector, an object of class \emph{rebase} (a matrix)
for an object of class \emph{rebase}, if \emph{rebase} is ommited return
a vector, otherwise a new object of class \emph{rebase} on the provided base.
}
\description{
Make a representation of a numeric vector on another,
possibly non integer, numeric base (radix)
}
\details{
Note that precision may be lost when converting from one base to another
and then back.
}
\examples{
x <- rebase(c(0,100), base = 10)
print(x)
rebase(x)
y <- rebase(c(0.8,90), 2, prec = 10)
print(y)
rebase(y) # 0.75, 90 (precision loss)
fi = (1 + sqrt(5))/2
rebase(8788, base=fi, prec=10)

}
